Add execution delay.
setExecutionDelay(long minDelay, long maxDelay) will cause the call to
the execute() method to be skipped when subscriptions show changes for
up to maxDelay milliseconds. If no subscription changes happen for
minDelay milliseconds, then execute() is called.

minDelay should be great enough to accumulate all the information
needed for a execution run, but not so great as to waste time waiting.

maxDelay should be set to insure that the plugin executes in the face
of a constant barrage of inputs.