Added the new "NodeIdentifier" and modified Node to be a MessageTransportClient and register with the MessageTransport.
Created a new "NodeMessage" and a single example, "AddClustersMessage",
and modified Node to receive this message and load the listed clusters.
All Clusters are now loaded using this API, similar to the way all
PlugIns are loaded by the Cluster using the AddPlugInMessage.  Also
cleaned up this loading code to accept fully-specified Cluster specs
using PropertyTrees.

Modified ClusterServesClusterManagement to let the Node see the
ClusterIdentifier, which simplifies some Node code.