Replaced the "ClusterINIParser" and "NodeINIParser" with a generic ComponentDescription parser.  The INI file format is now:
  # ignored comments
  [ignored section headers]
  insertionPoint = classname
  insertionPoint = classname(arg0, arg1, ..)

See the INIParser javadocs for further details, especially the
backwards compatibility for "plugin = " and "cluster = ".  In
particular, this should be transparent for existing INI files.

Node has been modified to accept these ComponentDescriptions.

Currently Node only loads Agents for itself and passes Plugin-level
descriptions to the PluginManager (which allows BinderFactories).