Remove remaining ClusterImpl dependencies in PluginManager. This should allow us to better hook up Agent and PluginManager.
Remove pass thru's for UIDService and LDMService related items.
PluginAdapter now uses DomainService for LDM Factories,
PrototypeRegistryService for prototype provider LDM services and the
UIDService.

Remaining getLDM calls are passed thru to stop-gap LDMService which
delegates to the DomainService and PrototypeRegistryService in
clusterimpl. It would be nice to get rid of this all together ... as
well as older UIDServer dependencies.