* Persistence.java: * FilePersistence.java: * DummyPersistence.java: * DatabasePersistence.java: * BasePersistence.java: Add suffix argument to disableWrite and readSequenceNumbers methods to allow arbitrary archive to be specified from the command line. In the case of DatabasePersistence, the suffix is parsed into its numeric value and used to query for the correct archive delta.
* Blackboard.java:
Add system property org.cougaar.core.cluster.persistence.sequence
used when org.cougaar.core.cluster.persistence.disableWrite is
true. It specifies the archive number to be used in rehydrating
the society. The form of the value is "_nnnnn" which is the suffix
on the sequence file to be used. When executing in really time, the
relevant suffix is the one on the sequence file having the date
desired. If running in non-real time. The correspondence must be
established manually.