Fixed mobility state-capture bug, where Domain instances were used instead of component-descriptions.  Upon mobility restart the domains were not reloaded.
Domains are now loaded via component-descriptions.  This is not
quite enough to make Domains INI-loadable, but it's close...

The Domain base-class, "DomainAdapter", now supports a component
parameter, which should be a List containing a single String entry
for the domain's name.

All this should be transparent to the plugins -- it should
"just work" or break horribly.  I've tested and it seems fine.