* CompletionTargetPlugin.java: Remove GLM-specific specialization (verbs to ignore) and provide interface whereby domain specializations can tailor the verbs to ignore.
* CompletionSourcePlugin.java:
Don't start waiting for completion until the relay targets have
stabilized. This makes it easier for subclass CompletionActions to
work reliably by not trying to invoke such actions until all
agents and nodes are present and accounted for.

* CompletionSocietyPlugin.java:
Remove GLM-specific specializations and provide CompletionAction
interface for extensions to specialize behavior such as publishing
the oplan or sending GLS.

* CompletionTargetPlugin.java:
A plugin for sensing agent activity and sending the completion
status of the agent as responses to CompletionRelays. Both the
presence and status of activity are sensed. If any Task or
PlanElement is added, removed, or changed, the cpuConsumption
countdown is restarted. cpuConsumption decreases linearly in time
from 1.0 to 0.0 over a ten minute span. taskCompletion is computed
as the average of the confidence ratings of the estimated
allocation result of all tasks excluding certain tasks for which
plugins fail to provide reliable confidence ratings
(DetermineRequirements, getLogSupport and MaintainInventory).
Updated values are sent as responses as warranted. Care is taken
to limit the frequency of interesting, but insignificant changes.
In particular, it really only matters if the laggard thresholds
are crossed. Otherwise, the changes are not significant so updates
are scheduled over long intervals.

* CompletionSourcePlugin.java:
Base class for plugins that send relays (CompletionNodePlugin and
CompletionSocietyPlugin). Ascertains the correct set of targets,
fashions a CompletionRelay and sends it. Retrieves the relay
responses and dispatches them to subclass methods for handling.

* CompletionSocietyPlugin.java:
The root completion plugin. Gathers completion status from all
nodes and decides when to advance execution time.