* org/cougaar/core/service/community/CommunityService.java: Bug #1636 -- The expansion of ABAs fails to track community membership Add methods to add and remove community change listeners.
* org/cougaar/core/service/community/CommunityChangeListener.java:
Bug #1636 -- The expansion of ABAs fails to track community membership
Interface to be implemented by community change listeners.

* org/cougaar/core/service/community/CommunityChangeEvent.java:
Bug #1636 -- The expansion of ABAs fails to track community membership
Event to send to community change listeners to describe the change.

* org/cougaar/core/service/community/CommunityChangeAdapter.java:
Bug #1636 -- The expansion of ABAs fails to track community membership
Adapter for building CommunityChangeListeners.