All basic component classes now have a public constant INSERTION_POINT, defined relative to its parent. Use this constant rather than the literal where possible. Use it for Node, Agent, AgentManager, PluginManager, Blackboard, DomainManager, and Plugin (in this case defined in PluginBase, all other cases defined in the so-named class/interface).
Fix bug 2482:
InitializerService.getComponentDescriptions should be given the
_parent_ insertion point, and it will return only items _below_ that
point. That is how the DBInitService works. File is different, since
it is just reading pre-split files. Arguably INIParser.parse or
FileInitializerServiceProvider should be modified so behavior matches,
but I have not done that here.
If users want items of a particular insertion point, specify the
parent, and then use
ComponentDescriptions.extractInsertionPointComponent(<insertion
point>)

Add a little debug/info level logging on loading of components.