Bug 1571: ComponentPlugin is now a trivial extension of BlackboardClientComponent.
One significant change:
getBindingSite returns a generic BindingSite, not a
PluginBindingSite. It _may_ however be cast to a PluginBindingSite,
which is what it will be.

However, users should have no need to make that cast.
The primary reason subclasses make that cast is to get the
AgentIdentifier, which may be retrieved directly: juse call
getAgentIdentifier(), (which uses the AgentIdentificationService).