Bug 1611
Made backwards compatible configuration initialization changes.
New -D arg
org.cougaar.core.node.InitializationComponent
Should specify the class of a component that can provide the
ComponentInitializerService
Shorthand is File for INI files, DB for CSMART database, and XML for
XML files.

The DBInitializerService is no longer always available, but only
available if the DBComponentInitializerServiceComponent is used, and
then only if, as before, the experiment.id -D arg was provided
(Correspondingly removed the DBInitializerServiceComponent which
provided it)


Removed the old ComponentInitializerServiceComponent which did a
branch, and replaced with XML, DB, and File versions to create the
appropriate ComponentInitializerServiceProvider, to provide the
appropriate ComponentInitializerService

This new -D arg is now used by the planning domain's
AssetInitializerService, and the community domain's
CommunityInitializerService, to decide how to initialize that data.
This new -D arg is now also provided by CSMART, to specify loading
from the CSMART DB.


Note the impact on bug 2522:
Although NodeAgent still has the ifPlanning and ifCommunity ugliness,
a user could provide their own AssetInitializerServiceComponent or
CommunityInitializerServiceComponent at BINDER priority, and the
default ones would not be used.