New WP bootstrap support.
The default configuration uses the old NS just like before.

Added node system properties to enable/disable the new WP/NS:
  org.cougaar.core.load.ns=BOOLEAN   (default is true)
  org.cougaar.core.load.wp=STRING    (default is true)
The above STRING can be set to "old", "new", or "both".  "true" is
equivalent to "old", which loads the old NS-based WP.  To enable
the new WP, set it to "both".