Bug 3328: If a Subscriber (incl any instance of BlackboardService) has no subscriptions, and opens a transaction while an outbox is being distributed, then the subscriber would be given the envelope. This includes servlets that use a Bboard service to publish something or do a query. It includes the OrgDataPlugin (which does its work in the load() method).
The result, if the timing is right, is that the PI holds on to the
envelope forever -- holding memory, and also blocking quiescence.

The fix is to require that the PI both have an open transaction, and
have at least 1 watcher -- that is, dont give the Subscriber the
envelope if it wont notice that you gave it to it.

Note that the checkin currently logs a WARN when the Fix is
exercised. I want to make sure I understand all the cases where it
gets excercised to be sure it has no ill effects. Then I will turn it
down to Info.