Bug 3359: Multiple changes to more closely track when the Distributor can accurately report on the quiescence of an agent -- tuning when the agent is considered active & local in the Node for quiescence purposes, particularly when moving agents.
Distributor
- Use new QuiescenceReportForDistributorService, not old QRS
- Every time a subscriber discards its rehydration info, if we have
not yet enabled the QRS, remove this subscriber from the list of
quiescence-relevant subscribers we are still expecting to rehydrate
- When rehydrating the distributor, collect a list of subscriber
states for subscribers that are quiescence-relevant. All such
subscribers are expected to create their subscriptions and have a
transaction early if not immediately, and to claim their rehydration
information. The agent will not be considered for local quiescence
(and hence will presumably keep the society from being quiescent)
until all such subscribers have had a transaction. So in particular,
servlets which open transactions only on user action are typically ignored for quiescence
to avoid delays.
- When stopping the distributor and releasing the QRS, first set this
agent Disabled for quiescence locally
- Clean up DEBUG log message in distribute -- the name of the first
subscriber that got something on its INBOX is now only printed at
DETAIL level, to avoid confusion.
- In distribute: Only tell the QM whether subscribers are
- In distribute: If we have not yet enabled QRS, and there are no more
q-relevant subscribers to rehydrate, then enable the QRS - after any
potential setting of current quiescence state

QuiescenceMonitor
- Store the last quiescence state of the distributor when persisting /
moving an agent
- New method to use a String to see if a Blackboard client (or
SubcriberStateKey) is quiescence-relevant

Persistence
BlackboardPersistence
- New method to get the SubscriberStateKeys -- used by Distributor to
see who has state they still need to rehydrate

QuiescenceReportServiceProvider
- split out QuiescenceState class
- code reorganization
- use FilteredIterator of agents in the agent container that say
.isEnabled() on their QuiescenceState for collecting agent states to
consider for quiescence -- instead of just the MessageAddress of the
agents in the container
- embed use of this iterator in the various checking methods
- clean up logic of announce Nonquiescence to avoid subtle errors
- only let a QRS client canceling quiescence cause Node to go non-q if
the agent for that client isEnabled (prevents StratTransProjPI
toggling quiescence on agent load)

QuiescenceState
- make private inner class in separate file. This class contains
quiescence state of an agent

NodeAgent
- Add in new QuiescenceReportForDistributorService
- No longer need to tell the QRS when an agent has been removed --
Distributor does that

QuiescenceReportForDistributorService
- Simple extenion of QuiescenReportService that allows the Distributor
to tell the QRS when to consider the local agent "Enabled" to count
for/against Node-level quiescence