Bugs 3603 and 3547
3547:
synchronize a couple more methods in
QuiescenceReportServiceProvider. In particular, when agents load they
do getService which adds a new entry to the map, from which we were
getting a ConcurrentModException

3603:
When an agent is duplicated elsewhere in the society for some reason,
the old copy gets isolated by the MTS, and agents talk only with the
new copy. Until we get Agent Suicide, the problem is that the old
agent is still counted by the Nodes quiescence service as local - so
the service waits for the dead agent to be quiescent before deciding that the node is
quiescent, and matches message numbers with this dead-and-out-of-date
agent. This will likely prevent quiescence.

Added a new servlet at /agentQuiescenceState to be loaded in every
Node, which shows the quiescence status of each Agent on the Node (and
the Node itself).
It also has a button for each Node by which you can mark the Agent as
"Dead".
This is not un-doable.
This has the effect of hiding that agent from the Nodes Quiescence
calculations altogether, as desired.
As noted, this whole mechanism should be replaced when we have Agent
Suicide.