Fixed yet another thread deadlock
Spurious wakeup could happen while thread is in the process of being given out by the thread pool.

Made the schedulable=null be the definative test  for spurious wakeup and do not allow scheduleable to be set unless the thread is in the wait state.

Seperated the use of variables in_use and schedulable
 When schedulable is set, there is work to be done and the thread had to have been woken by a notify (I hope)
when in_use is false the thread can allocated to a scheduable, but the schedulable still has get the RunLock, which is only available when the thread is waiting.